split component and bug fix
